+++ redash/handlers/base.py
@@ -117,7 +117,7 @@ def paginate(query_set, page, page_size, serializer, **kwargs):
 
 def org_scoped_rule(rule):
     if settings.MULTI_ORG:
-        return "/<org_slug:org_slug>{}".format(rule)
+        return "/<org_slug>{}".format(rule)
 
     return rule
