+++ redash/worker.py
@@ -8,8 +8,10 @@ from flask import current_app
 from celery import Celery
 from celery.schedules import crontab
 from celery.signals import worker_process_init
+
 from redash import create_app, settings
-from redash.metrics import celery as celery_metrics
+from redash.metrics import celery as celery_metrics  # noqa
+
 
 celery = Celery('redash',
                 broker=settings.CELERY_BROKER,
